+++ src/app/search/FilterHelp.tsx
@@ -546,8 +546,9 @@ function FilterHelp({ destinyVersion }: { destinyVersion: 1 | 2 }) {
                   <span>source:zavala,source:strikes</span> <span>source:ikora</span>
                   <span>source:drifter,source:gambit</span> <span>source:ep</span>
                   <span>source:raid</span> <span>source:leviathan</span> <span>source:sos</span>
-                  <span>source:eow</span> <span>source:lastwish</span> <span>source:prestige</span>
-                  <span>source:adventure</span>
+                  <span>source:eow</span> <span>source:lastwish</span> <span>source:scourge</span>
+                  <span>source:prestige</span> <span>source:adventure</span>{' '}
+                  <span>source:blackarmory</span>
                 </td>
                 <td>
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.FWC') }} />
@@ -577,8 +578,10 @@ function FilterHelp({ destinyVersion }: { destinyVersion: 1 | 2 }) {
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.SoS') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.EoW') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.LastWish') }} />
+                  <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.Scourge') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.Prestige') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.Adventure') }} />
+                  <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.BlackArmory') }} />
                 </td>
               </tr>
             )}